Versão "final" do script para extrair necessidade de importação de DUEs
1 - Pega acessos veículo do período e recintos, se passado.
2 - Em seguida pega escaneamentos correspondentes (efetuados logo após no fluxo)
3 - Recupera os _ids das imagens de escaneamento adquiridas do MongoDB. Obs.: o centro da pesquisa é sempre em torno do número do contêiner.
4 - Recuperar do Receita Data as DUEs e ItensDUE a partir da lista de contêineres
5 - Atualizar metadata do Mongo por _id e Acesso por id com número da DUE, a partir da lista do RD
6 - Atualizar ou inserir DUEs a partir do exportado pelo RD
7 - Inserir nomes das empresas na tabela

+++ virasana/scripts/lista_conteineres_exportacao_sem_dues.py
@@ -5,15 +5,16 @@ from datetime import datetime, timedelta, time
 import pandas as pd
 
 sys.path.append('.')
-from virasana.integracao.due.due_alchemy import Due, DueItem
-from virasana.integracao.due.manager_conteineres_dues import get_conteineres_escaneados_sem_due
+from virasana.integracao.bagagens.bagagens_risco import importa_cnpjs
+from virasana.integracao.due.manager_conteineres_dues import (get_conteineres_escaneados_sem_due,
+                                                              set_conteineres_escaneados_sem_due,
+                                                              integra_dues, integra_dues_itens)
 from virasana.integracao.due.rd_baixa_dues_novas import dues_rd
 
 
-def exporta_ctrs(session):
-    # Obtém a data de ontem
-    lista_final = get_conteineres_escaneados_sem_due(session, datainicio=inicio_ontem,
-                                                     datafim=fim_ontem, codigoRecinto='8931359')
+def exporta_ctrs(session, inicio: datetime, fim: datetime, codigos_recintos: list):
+    lista_final = get_conteineres_escaneados_sem_due(session, inicio,
+                                                     fim, codigos_recintos)
     df = pd.DataFrame(lista_final[1:], columns=lista_final[0])
     df.to_csv('escaneamentos_sem_due.csv', index=False)
 
@@ -30,60 +31,63 @@ def insert_from_dataframe(session, classe, df):
     finally:
         session.close()
 
+def atualiza_acesso_e_mongo(db, session):
+    # Passo 5: Atualizar metadata do Mongo por _id e Acesso por id com número da DUE
+    # Não tratar erro, se houver, interromper aqui porque senão o passo mais importante ficará para trás
+    # Se der erro, operador precisa tratar e identificar, porque esta amarração é a parte mais importante
+    df_dues = pd.read_csv('dues.csv')
+    df_escaneamentos_sem_due = pd.read_csv('escaneamentos_sem_due.csv')
+    set_conteineres_escaneados_sem_due(db, session, df_escaneamentos_sem_due, df_dues)
 
-def importa_dues():
-    # Passo 5: Inserir DUEs e itensDUE a partir dos dfs acima
-    # Passo 6: Atualizar metadata e Acesso com número da DUE* (TODO)
-    # Passo 7: Inserir nomes das empresas na tabela XXXXXX (TODO)
-    # *Falta gerar o csv da ligação do contêiner com a DUE na parte do RD
+
+def importa_dues(session):
     if not os.path.exists('dues.csv'):
         print('Pulando dues - arquivo não existe')
     else:
+        # Passo 6: Inserir DUEs a partir do df
         df_dues = pd.read_csv('dues.csv')
-        df_dues['ni_declarante'] = df_dues['ni_declarante'].astype(str).str[:14].str.zfill(14)
-        df_dues['cnpj_estabelecimento_exportador'] = (
-            df_dues['cnpj_estabelecimento_exportador'].astype(str).str[:14].str.zfill(14))
-        insert_from_dataframe(session, Due, df_dues)
-        os.remove('dues.csv')
+        df_dues = df_dues.fillna('').drop_duplicates()
+        if integra_dues(session, df_dues):
+            os.remove('dues.csv')
 
+    # Passo 6b: Inserir DUEItens a partir do df
     if not os.path.exists('itens_dues.csv'):
         print('Pulando itens_dues - arquivo não existe')
     else:
         df_itens_dues = pd.read_csv('itens_dues.csv')
-        insert_from_dataframe(session, DueItem, df_itens_dues)
-        os.remove('itens_dues.csv')
-
-    if not os.path.exists('cnpjs_nomes.csv'):
-        print('Pulando cnpjs_nomes.csv - arquivo não existe')
-    else:
-        df_cnpjs = pd.read_csv('cnpjs_nomes.csv')
-        os.remove('cnpjs_nomes.csv')
+        df_itens_dues = df_itens_dues.fillna('').drop_duplicates()
+        if integra_dues_itens(session, df_itens_dues):
+            os.remove('itens_dues.csv')
 
 
 if __name__ == '__main__':  # pragma: no-cover
     from ajna_commons.flask.conf import SQL_URI
     from sqlalchemy import create_engine
     from sqlalchemy.orm import sessionmaker
+    from virasana.db import mongodb
 
     engine = create_engine(SQL_URI)
     Session = sessionmaker(bind=engine)
     session = Session()
 
-    ontem = datetime.now() - timedelta(days=1)
-    # Define o início e o fim do dia de ontem
-    inicio_ontem = datetime.combine(ontem, time.min)
-    fim_ontem = datetime.combine(ontem, time.max)
+    # Define o fim às 00h de hoje e início 5 dias antes
+    fim = datetime.combine(datetime.now(), time.min)
+    inicio = fim - timedelta(days=5)
+
     operacao = input('Qual a operação desejada? (E)exportar lista, (I)mportar DUEs, (R)eceita Data?')
 
     if operacao == 'E':
-        exporta_ctrs()
+        # Passo 1, 2 e 3
+        exporta_ctrs(session, inicio, fim, ['8931359'])
     elif operacao == 'I':
-        importa_dues()
+        # Passo 5
+        atualiza_acesso_e_mongo(mongodb, session)
+        # Passo 6
+        importa_dues(session)
+        # Passo 7: Inserir nomes das empresas na tabela
+        importa_cnpjs(session, 'cnpjs_nomes.csv')
     elif operacao == 'R':
         # Passo 4: Recuperar do Receita Data CÓDIGO FAKE DE EXEMPLO!!!
         # Abaixo apenas exemplo, hoje é preciso rodar no Jupyter Notebook do RD na mão
         df = pd.read_csv('escaneamentos_sem_due.csv')
-        dues_rd(inicio_ontem, fim_ontem, df['numero_conteiner'].unique(), None)
-        df_dues_conteiner = pd.read_csv('dues_conteiner.csv')
-        merge_df = df.merge(df_dues_conteiner, left_on='numero_container', right_on='id_conteiner', how='inner')
-        merge_df.to_csv('escaneamentos_sem_due_atualizado.csv', index=False)
+        dues_rd(None, inicio, fim, df['numero_conteiner'].unique(), ['8931359'])
